server/requirements.txt
asgiref==3.6.0
channels==4.0.0
Django==4.2
numpy==1.22.0
opencv_contrib_python==4.7.0.68
opencv_python==4.7.0.68
opencv_python_headless==4.8.1.78
paho_mqtt==1.6.1
celery==5.3.6